refactor(app): NAV_ITEMS import 수정
